Used docstrings  and improved some type hints using the Typing module

+++ Python_BackEnd/Products/Models/Products.py
@@ -1,5 +1,6 @@
 import uuid
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
@@ -8,4 +9,4 @@ class Products:
     title: str = ""
     imagePath: str = ""
     retailer: str = ""
-    cost: float = 0.0
+    cost: Decimal = 0.0
